Move logging API to util package
